+++ sdk/formrecognizer/azure-ai-formrecognizer/CHANGELOG.md
@@ -19,6 +19,8 @@
 - For recognize receipt methods, parameters have been renamed to `receipt` and `receipt_url`.
 - `created_on` and `last_modified` are renamed to `requested_on` and `completed_on` in the
 `CustomFormModel`  and `CustomFormModelInfo` models.
+- `models` property of `CustomFormModel` is renamed to `submodels`
+- `CustomFormSubModel` is renamed to `CustomFormSubmodel`
 
 **New features**